Names of uploaded files can be seen. UI is the next step
